requirements.txt
playwright
beautifulsoup4
aiohttp
python-dotenv
